emacs: add EIN
This patch adds the EIN package, as well as some other customization.

EIN doesn't really fit into the workflow of a single Jupyter daemon which is always open, so we
added my/open-jupyter-notebook to help with that. It reads the URL of the Jupyter notebook from the
customization file, log in if necessary and then open the notebook.

We also set some of its faces in order to better fit the Doom theme.

In addition, we disable Emacs's auto completion, as it behaves strangely in EIN. Completion is done
either by typing the . character after an object, or by pressing C-c TAB.