Neovim
This patch makes a small fix for the vim configuration in order to work in Neovim. It also makes the
shell prefer Neovim as the alternate editor if it is present.

Neovim expects its configuration file to be at ~/.config/nvim/init.vim, which is why the init.vim
symlink has been committed in this patch.

As explained here, cryptmethod isn't implemented in Neovim
https://github.com/neovim/neovim/issues/701